holy shit its actually working code for once (TPBar fill)

+++ VerticalSlice2D/Assets/Scripts/Player.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Player : MonoBehaviour
 {
-    public double TP = 0;
+    public float TP = 0;
 
     // Start is called before the first frame update
     void Start()
